+++ pubspec.yaml
@@ -41,6 +41,7 @@ dependencies:
   dio: ^5.9.0
   flutter_dotenv: ^5.0.2
   flutter_screenutil: ^5.9.3
+  youtube_player_flutter: ^9.1.2
 
 dev_dependencies:
   flutter_test:
